добавил сцены, надо разобраться почему горит play, как будто я на неё навёлся

+++ src/core/GameActor.cpp
@@ -2,12 +2,12 @@
 
 
 
-GameActor::GameActor(float x, float y) : x(x), y(y) {}
+GameActor::GameActor(float x, float y) : x(static_cast<int>(x)), y(static_cast<int>(y)) {}
 
 
 GameActor::~GameActor() {}
 
-void GameActor::AMove() { x -= 1.0f; }
-void GameActor::SMove() { y -= 1.0f; }
-void GameActor::DMove() { x += 1.0f; }
-void GameActor::WMove() { y += 1.0f; }
+void GameActor::AMove() { x -= 1; }
+void GameActor::SMove() { y -= 1; }
+void GameActor::DMove() { x += 1; }
+void GameActor::WMove() { y += 1; }
